Address cache routing review feedback
