+++ crates/mesh-llm-host-runtime/src/network/openai/response.rs
@@ -1,3 +1,4 @@
+mod cache_cost;
 mod cancellation;
 mod common;
 mod dispatch;
@@ -11,8 +12,9 @@ mod routing;
 mod send;
 mod stream_translation;
 
+pub(super) use cache_cost::CacheCostObservation;
 pub(super) use common::{
-    CacheCostObservation, ResponseRetryPolicy, RouteAttemptLoggingContext, RouteAttemptResult,
+    ResponseRetryPolicy, RouteAttemptLoggingContext, RouteAttemptResult,
     attempt_outcome_for_result, completion_tokens_for_result, parse_token_usage_from_json_body,
     request_outcome_for_status, request_service_for_target, route_attempt_result_label,
     target_health_outcome_for_attempt,
