+++ crates/mesh-llm-host-runtime/src/network/openai/response/json_adaptation.rs
@@ -1,6 +1,7 @@
+use super::cache_cost::parse_cache_cost_from_json_body;
 use super::common::{
-    ResponseRetryPolicy, RouteAttemptResult, parse_cache_cost_from_json_body,
-    parse_token_usage_from_json_body, retryable_quality_result,
+    ResponseRetryPolicy, RouteAttemptResult, parse_token_usage_from_json_body,
+    retryable_quality_result,
 };
 use super::probe::{
     ResponseBodyReadLimits, ResponseProbe, read_transformed_response_body,
